requirements.txt
appdirs==1.4.4
APScheduler==3.7.0
bidict==0.21.2
click==7.1.2
distlib==0.3.1
dnspython==1.16.0
eventlet==0.30.2
filelock==3.0.12
greenlet==1.0.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pbr==5.5.1
pydealer==1.4.0
python-engineio==4.0.1
python-socketio==5.1.0
pytz==2021.1
six==1.15.0
stevedore==3.3.0
tzlocal==2.1
virtualenv==20.4.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
